Fix parsing of Decimal types
Decimal types require the schema to determine precision, scale and what base type the data is.
For some reason it is passed in as null. I looked at all other implementations of `ReadSingle` and
cannot see any reason why it has to be null. So I passed in the SchemaElement.