`Update NullableTimeSpan assignment in TimeSpan_Simplest_Serde`
Updated the `TimeSpan_Simplest_Serde()` method in the `Parquet.Test.Serialisation` namespace in the `ParquetSerializerTest.cs` file. The `NullableTimeSpan` property in the `TimespanRecord` object, which was previously always set to `null`, is now conditionally set to `TimeSpan.Parse("01:00:03")` if the index `i` is an even number, otherwise it remains `null`.