Pass property attributes down to list data field
Attributes decorating a list-type class member are not accessible to
ConstructDataField when creating the element data field.  This commit
passes the ClassMember object representing the list down to
ConstructListField so that it can pass it to MakeField for the list
element.  This allows decorating a list of strings with, for example,
ParquetRequired, so that the list elements are not nullable.  Or a
list of DateTimes as ParquetTimestamp, to control how the DateTime
elements are serialized.